requirements.txt
google-cloud-pubsub==2.12.0